backend/requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
SQLAlchemy==1.4.48  
boto3==1.26.159
alembic==1.11.1
aiohttp-devtools
pyyaml
aiosqlite
aiohttp-sqlalchemy==0.34.0
aiohttp-swagger
aioauth-client
aiohttp_session[secure]
aiohttp_security
aiohttp_jwt

# aiomysql==0.0.22  # if using MySQL
# or
# aiopg==1.3.0  # if using PostgreSQL
